requirements.txt
aiodns==3.2.0
aiohttp==3.11.14
atomicwrites-homeassistant==1.4.1
attrs==25.3.0
awesomeversion==24.6.0
blockbuster==1.5.24
brotli==1.1.0
ciso8601==2.3.2
colorlog==6.9.0
cpe==1.3.1
cryptography==44.0.2
debugpy==1.8.13
deepmerge==2.0
dirhash==0.5.0
docker==7.1.0
faust-cchardet==2.1.19
gitpython==3.1.44
jinja2==3.1.6
log-rate-limit==1.4.2
orjson==3.10.16
pulsectl==24.12.0
pyudev==0.24.3
PyYAML==6.0.2
requests==2.32.3
securetar==2025.2.1
sentry-sdk==2.24.1
setuptools==78.1.0
voluptuous==0.15.2
dbus-fast==2.43.0
typing_extensions==4.12.2
zlib-fast==0.2.1
